Move seeding to seed.js. Start to build basic api response in server
