+++ client/index.jsx
@@ -1,16 +1,27 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/App.jsx';
+import axios from 'axios';
 
-var listing = {
-  maxGuests: 5,
-  maxInfants: 3,
-  chargePerNight: 150,
-  cleaningFee: 10,
-  serviceFee: 10,
-  occupancyFee: 10,
-  rating: 9,
-  numberOfRatings: 300,
-};
+axios.get('/api/dblistings', {
+  params: {
+    listing: 1,
+  }
+})
+  .then((listing) => {
+    ReactDOM.render(<App listing={listing}/>, document.getElementById('root'));
+  })
+  .catch((error) => {
+    console.log('error fetching listing data', error);
+  });
 
-ReactDOM.render(<App listing={listing}/>, document.getElementById('root'));
+// var listing = {
+//   maxGuests: 5,
+//   maxInfants: 3,
+//   chargePerNight: 150,
+//   cleaningFee: 10,
+//   serviceFee: 10,
+//   occupancyFee: 10,
+//   rating: 9,
+//   numberOfRatings: 300,
+// };
