+++ server/server.js
@@ -9,3 +9,7 @@ app.use(morgan('tiny'));
 app.use(express.static('public'));
 
 app.listen(port);
+
+app.get('/api/dblistings', (req, res) => {
+
+});
